Split brush world presentation
